requirements.txt
future
emoji
lxml
beautifulsoup4
requests
sqlalchemy==1.3.20
python-telegram-bot==12.8
envparse
faker
psycopg2-binary
feedparser
pynewtonmath
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
googletrans
jikanpy
speedtest-cli
coffeehouse
pymongo
regex
bleach
git+https://github.com/starry69/python-markdown2.git
wikipedia
telethon==1.16.4
PyYAML>5.0
spamwatch
alphabet_detector
loguru
pyrate-limiter
gtts
nekos.py
fontTools
typing
tswift
cachetools
pyowm
tabulate
parsel
googletrans
mtranslate
pyrogram
tgcrypto
configparser
pytz
hachoir
pybase64
pySmartDL
validators
youtube_dl
google_trans_new
#googletrans
pyrogram
bing_image_downloader
html2text
bs4
motor
search_engine_parser
dnspython
youtube_search
telegraph
PyDictionary==2.0.1
youtube-search-python
secureme
